Tolerate universal tag number zero.
We currently reject the zero byte as a tag deep in parseTag. This makes
handling EOC easier, as EOC as encoded as an empty pseudo-element. ASN.1
makes this work by reserving universal tag zero for encodings.

This, unfortunately, leaves it ambiguous parsing an element without an
expected tag should work. If field has type ANY, and one sees {0x00,
0x00}, should that parse? What if it's non-empty, such as {0x00, 0x01,
0x00}? What if it has tag number zero, but the constructed bit is set,
i.e. {0x20, 0x00}?

A lot of parsers are a little lax here, which means that they'll accept
inputs that der2ascii cannot print. This CL instead removes EOC
awareness from the low-level element parsers. Instead, this is punted to
the two parseElement call sites, which already need to specially handle
EOC:

When decoding elements, we'll simply treat any unexpected EOC as an
element. So if a definite-length element contains `0000`, that will no
be printed as [UNIVERSAL 0 PRIMITIVE] {}, rather than terminating
decode. If we are parsing elements of an indefinite-length element,
however, EOCs will finish the element as before. In either context,
non-empty zero elements are elements as before.

Note this does mean that a string of trailing zeros in the input will
now get decoded as a bunch of elements rather than all zeros. If that
becomes annoying, we can probably treat the outermost calls special.

An unexpected EOC, however, still terminates isMadeOfElements. This is
so the we do not interpret OCTET STRING { `0000` } as containing
elements. Require that input to be more well-formed-looking before
recursing into primitives.