Add bit string literals for ease of encoding BIT STRINGs
BIT STRING is distinct from OCTET STRING in that it has a leading byte
specifying the number of "unimportant" bits in the last octet. This adds
syntax (b`...`) similar to hex literals, which generates this prefix in
a convenient fashion. See language.txt for details.

Includes changes to ascii2der, as well as a changing the existing
der2ascii heuristic for well-formed BIT STRINGs to use the new syntax
instead.