Add support for RELATIVE-OID
Mirroring the 1.2.3 syntax for OIDs, a relative OID is .1.2.3. Note this
does mean .1 will be read as a relative OID, but 0.1 was already read as
an OID. If we ever feel the need to support ASN.1's weird floating point
bits, we'll probably need to do something more complicated anyway.

This seems to have some precedent in at least the examples of RFC 9090.
Interestingly, https://oid-base.com/faq.htm#mib cites the opposite
convention.